requirements.txt
xlsxwriter>=3.1
openai>=1.0
python-dotenv>=1.0
